Rework routes

+++ src/routes/accomplishment/controller.ts
@@ -5,6 +5,10 @@ export async function getAccomplishment(
   fastify: FastifyInstance,
   accomplishmentId: number
 ) {
+  if (!accomplishmentId) {
+    throw fastify.httpErrors.badRequest("Invalid accomplishment id");
+  }
+
   const accomplishment = await fastify.prisma.accomplishment.getAccomplishment(
     accomplishmentId
   );
@@ -18,7 +22,7 @@ export async function getAccomplishment(
 
 export async function getAccomplishments(fastify: FastifyInstance) {
   const accomplishments =
-    await fastify.prisma.accomplishment.getAccomplishments();
+    await fastify.prisma.accomplishment.getManyAccomplishment();
 
   if (!accomplishments || !accomplishments.length) {
     throw fastify.httpErrors.notFound("No Accomplishments in DB");
@@ -29,9 +33,27 @@ export async function getAccomplishments(fastify: FastifyInstance) {
 
 export async function createAccomplishment(
   fastify: FastifyInstance,
-  accomplishmentInfo: AccomplishmentInfo
+  accomplishmentInfo: AccomplishmentInfo,
+  userId: number,
+  challengeId: number
 ) {
-  await fastify.prisma.accomplishment.createAccomplishment(accomplishmentInfo);
+  if (!userId) {
+    throw fastify.httpErrors.badRequest("Invalid user id");
+  }
+
+  if (!challengeId) {
+    throw fastify.httpErrors.badRequest("Invalid challenge id");
+  }
+
+  if (!accomplishmentInfo) {
+    throw fastify.httpErrors.badRequest("No accomplishment info provided");
+  }
+
+  await fastify.prisma.accomplishment.createAccomplishment(
+    accomplishmentInfo,
+    userId,
+    challengeId
+  );
 }
 
 export async function updateAccomplishment(
@@ -39,9 +61,13 @@ export async function updateAccomplishment(
   accomplishmentInfo: AccomplishmentInfo,
   accomplishmentId: number
 ) {
+  if (!accomplishmentId) {
+    throw fastify.httpErrors.badRequest("Invalid accomplishment id");
+  }
+
   await fastify.prisma.accomplishment.updateAccomplishment(
-    accomplishmentInfo,
-    accomplishmentId
+    accomplishmentId,
+    accomplishmentInfo
   );
 }
 
@@ -49,20 +75,29 @@ export async function deleteAccomplishment(
   fastify: FastifyInstance,
   accomplishmentId: number
 ) {
+  if (!accomplishmentId) {
+    throw fastify.httpErrors.badRequest("Invalid accomplishment id");
+  }
+
   await fastify.prisma.accomplishment.deleteAccomplishment(accomplishmentId);
 }
 
 export async function validateAccomplishment(
   fastify: FastifyInstance,
-  state: 1 | -1,
+  validation: 1 | -1,
   accomplishmentId: number
 ) {
-  if (state < -1 || state > 1) {
+  if (!accomplishmentId) {
+    throw fastify.httpErrors.badRequest("Invalid accomplishment id");
+  }
+
+  if (!(validation === 1 || validation === -1)) {
     throw fastify.httpErrors.badRequest("Invalid validation state");
   }
 
-  await fastify.prisma.accomplishment.validateAccomplishment(
+  await fastify.prisma.accomplishment.updateAccomplishment(
     accomplishmentId,
-    state
+    {},
+    validation
   );
 }
