+++ src/routes/challenge/controller.ts
@@ -7,13 +7,22 @@ import {
 
 export async function createChallenge(
   fastify: FastifyInstance,
-  challengeInfo: ChallengeInfo
+  challengeInfo: ChallengeInfo,
+  creatorId: number
 ) {
+  if (!challengeInfo) {
+    throw fastify.httpErrors.badRequest("No challenge info provided");
+  }
+
   if (challengeInfo.reward && challengeInfo.reward < 0) {
     throw fastify.httpErrors.badRequest("Reward must pe positive");
   }
 
-  await fastify.prisma.challenge.createChallenge(challengeInfo);
+  if (!creatorId) {
+    throw fastify.httpErrors.badRequest("Invalid creator id");
+  }
+
+  await fastify.prisma.challenge.createChallenge(challengeInfo, creatorId);
 }
 
 export async function updateChallenge(
@@ -21,8 +30,16 @@ export async function updateChallenge(
   challengeId: number,
   challengeInfo: ChallengeInfo
 ) {
+  if (!challengeInfo) {
+    throw fastify.httpErrors.badRequest("No challenge info provided");
+  }
+
   if (challengeInfo.reward && challengeInfo.reward < 0) {
     throw fastify.httpErrors.badRequest("Reward must pe positive");
+  }
+
+  if (!challengeId) {
+    throw fastify.httpErrors.badRequest("Invalid challenge id");
   }
 
   await fastify.prisma.challenge.updateChallenge(challengeInfo, challengeId);
@@ -32,6 +49,10 @@ export async function deleteChallenge(
   fastify: FastifyInstance,
   challengeId: number
 ) {
+  if (!challengeId) {
+    throw fastify.httpErrors.badRequest("Invalid challenge id");
+  }
+
   await fastify.prisma.challenge.deleteChallenge(challengeId);
 }
 
@@ -39,6 +60,10 @@ export async function getChallenge(
   fastify: FastifyInstance,
   challengeId: number
 ) {
+  if (!challengeId) {
+    throw fastify.httpErrors.badRequest("Invalid challenge id");
+  }
+
   const challenge = await fastify.prisma.challenge.getChallenge(challengeId);
 
   //Check if challenge not found
@@ -50,7 +75,7 @@ export async function getChallenge(
 }
 
 export async function getChallenges(fastify: FastifyInstance) {
-  const challenges = await fastify.prisma.challenge.getChallenges();
+  const challenges = await fastify.prisma.challenge.getManyChallenge();
 
   //Check if challenge DB empty
   if (!challenges || !challenges.length) {
